VGG-Test/requirements.txt
tensorflow-macos
tensorflow-metal
opencv-contrib-python
pandas
matplotlib
seaborn
numpy
scikit-learn
jupyter
requests